Implement "str-of" pattern
A major re-write of the pattern matching logic to cleanly separate the
notion of a "str-of" representation of another type, and in some cases a
"num-of" representation.

After this, an Int always representation pure integers, but may be
wrapped in a StrRepr to indicate that in the source data the integers
are a string encoding of integers (with the StrRepr storing the pattern
used to encode the integers). Likewise, a string containing a number
which always represents a POSIX-style offset date can be represented
as a StrRepr(NumRepr(DateTime(...))).

At this point everything passes the test suite, but there's still plenty
of coverage work to do.