Implement a crude progress measure
Construct a set of all IDs in the loaded data and remove each ID from
the set as we attempt to match on it. Shift the analysis into a
background thread and show a regularly updated progress display from the
main thread indicating how "empty" the ID list is.

Also added an option to avoid tracking the progress (as the initial step
of constructing the set of all IDs is potentially lengthy and users may
prefer speed over regular updates if they trust structa).