Move output to XML/XSLT transforms
All type objects (and supporting classes) now support XML output, and
the CLI interface uses an XSLT transform to generate its output (with
some fancy color, courtesy of blessings). This also paves the way for a
nice web interface which can also use some XSLT to perform a vareity of
transforms with various parameters as configuration.