Remove popularity exclusion of bad values
The original assumption behind this feature was that, although not
technically "correct" it was a "reasonable" assumption that bad data
would always be in the minority.

However, it turns out (discovered while writing the docs!), there are
cases where, although still in the minority, a single bad value can
out-rank in popularity all other "good" values. Consider a data-set
containing strings of unique integers. Now consider that, in lieu of
certain values a well-meaning researcher has stuck in "?" several times.
The "?" values will have a count of n (>1), while all the other values
will have a count of 1. In this scenario, the bad values can be the most
popular, while still remaining below the "bad threshold".

Hence, with a certain amount of regret (it was a neat hack!) we have to
get rid of this.