Improved merging
A fairly major re-write of the equality, addition, and zipping code for
Dict and Tuple. Found an example of a large (~200MB) JSON struct that
crashed the prior merge code. The new algorithm treats as equal specific
fields that match a general type, and containers that match >= the
similarity threshold.

However, this in turn meant that the iterative cycle for successive
merges failed (all merged structs are by definition equal to their
unmerged counter-part). That required the addition of the "size"
property (providing the recursive count of Type derivatives under the
queried instance) to permit termination of the merge cycle.