Fix a special case of dictionaries merging
When dictionaries under-go merge at the end of analysis, it is possible
for one side of a dictionary addition to have fields, and the other side
to be a type (compatible with those fields' values). In this case, the
former code messes things up and fails to merge things correctly.