Fix rematch_sample (replaced with Redo)
It's not enough to re-run match when merging Fields to a Scalar type in
a Dict. Firstly, the redo point can be below the top level so a
recursive search is required to find any redo points after merging Dict
instances. Secondly, a full analyze() run is needed rather than match()
as the merged instances can potentially contain containers themselves.